apply refactorizations
