+++ .github/workflows/docker-publish.yaml
@@ -1,4 +1,4 @@
-name: Build and publish relayer image
+name: Build and publish container image
 
 on:
   push:
@@ -19,7 +19,7 @@ jobs:
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v3
 
-      - name: Login to GitHub Container Registry
+      - name: Log in to the container registry
         uses: docker/login-action@5e57cd118135c172c3672efd75eb46360885c0ef # v3.6.0
         with:
           registry: ghcr.io
